requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.5.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-filter==2.3.0
django-generic-follow==0.4.2
django-jet==1.0.8
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-autoviv==0.0.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
RestResponse==3.1.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
